I think some kind of work was in progress here, but I can't remember, and I'm going to bed. Just committing so I don't lose anything.

+++ app/src/main/java/Logging/AntiLogger.java
@@ -17,7 +17,7 @@ public class AntiLogger implements MasterLogger {
     }
 
     @Override
-    public void recordTrickStateAtBotTurn(double tricksPlayed, double tricksWon, Deck hand, Deck playedCards, Card teamPlayedCard, boolean teamWon, int playOrder) {
+    public void recordTrickStateAtBotTurn(int tricksPlayed, int tricksWon, Deck hand, Deck playedCards, Card teamPlayedCard, boolean teamWon, int playOrder) {
     }
 
     @Override
